[ui] Hide __ASSET_JOB references on the sensor pages
